+++ src/battle-scene.ts
@@ -11,7 +11,8 @@ import { Phase } from "./phase";
 import { initGameSpeed } from "./system/game-speed";
 import { Biome } from "./data/enums/biome";
 import { Arena, ArenaBase } from "./field/arena";
-import { GameData, PlayerGender } from "./system/game-data";
+import { GameData } from "./system/game-data";
+import { PlayerGender } from "./data/enums/player-gender";
 import { TextStyle, addTextObject } from "./ui/text";
 import { Moves } from "./data/enums/moves";
 import { allMoves } from "./data/move";
