+++ src/data/enums/game-data-type.ts
@@ -1,0 +1,10 @@
+/**
+ * enum for the game data types
+ */
+export enum GameDataType {
+  SYSTEM,
+  SESSION,
+  SETTINGS,
+  TUTORIALS,
+  SEEN_DIALOGUES
+}
