+++ src/data/enums/passive.ts
@@ -1,0 +1,7 @@
+/**
+ * enum for passive
+ */
+export enum Passive {
+  UNLOCKED = 1,
+  ENABLED = 2
+}
