+++ src/data/enums/player-gender.ts
@@ -1,0 +1,8 @@
+/**
+ * enum for the players gender
+ */
+export enum PlayerGender {
+  UNSET,
+  MALE,
+  FEMALE
+}
