+++ src/ui/summary-ui-handler.ts
@@ -17,7 +17,7 @@ import { StatusEffect } from "../data/status-effect";
 import { getBiomeName } from "../data/biomes";
 import { Nature, getNatureStatMultiplier } from "../data/nature";
 import { loggedInUser } from "../account";
-import { PlayerGender } from "../system/game-data";
+import { PlayerGender } from "#app/data/enums/player-gender";
 import { Variant, getVariantTint } from "#app/data/variant";
 import {Button} from "../enums/buttons";
 import { Ability } from "../data/ability.js";
